Bug #654389 modify converity of Test should be assignment & Structurally dead code & sizeof not portable & Result is not floating-point & Resource leak in object
[root cause ] converity failed
[changes ] modify converity case : 132620 & 132934 & 60168 & 132843 & 131844 & 131843 & 132323 & 171233 & 133344 & 132800 & 132797 & 132794 & 132785 & 132751 & 132750 & 132748 & 132747 & 132742 & 131730 & 131713 & 107766
[side effects] N/A
[self test ]<board, Test cases summary>:<9850ka, camera func ok>
[reviewers ] Miller.song
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ hal3_2v1/SprdCamera3OEMIf.cpp
@@ -3734,8 +3734,10 @@ void SprdCamera3OEMIf::receivePreviewFrame(struct camera_frame_type *frame) {
         int64_t timestamp = 0;
 
         fps_range_offset = 1000000000 / fps_param.max_fps;
-        fps_range_low = (1000000000 / fps_param.max_fps) * (1 - 0.015);
-        fps_range_up = (1000000000 / fps_param.min_fps) * (1 + 0.015);
+        fps_range_low =
+            (int64_t)((1000000000.0 / fps_param.max_fps) * (1 - 0.015));
+        fps_range_up =
+            (int64_t)((1000000000.0 / fps_param.min_fps) * (1 + 0.015));
 
         if (mPrvBufferTimestamp == 0) {
             buffer_timestamp_fps = frame->timestamp;
